# Optimise alignment/part1 native global alignment