--- a/include/seqan3/alignment/pairwise/align_pairwise.hpp
+++ b/include/seqan3/alignment/pairwise/align_pairwise.hpp
@@ -175,9 +175,11 @@ constexpr auto align_pairwise(sequence_t && sequences,
             return seqan3::seq;
     };
 
+    using alignment_result_t = typename traits_t::alignment_result_type;
     // Create a two-way executor for the alignment.
     detail::alignment_executor_two_way executor{std::move(seq_view),
                                                 std::move(algorithm),
+                                                alignment_result_t{},
                                                 traits_t::alignments_per_vector,
                                                 get_execution_rule()};
     // Return the range over the alignments.