… include/seqan3/alignment/pairwise/detail/pairwise_alignment_algorithm.hpp
… include/seqan3/alignment/pairwise/detail/policy_affine_gap_recursion.hpp
… include/seqan3/alignment/pairwise/edit_distance_algorithm.hpp
… include/seqan3/alignment/pairwise/edit_distance_fwd.hpp
… include/seqan3/alignment/pairwise/execution/alignment_executor_two_way.hpp
… test/performance/alignment/edit_distance_unbanded_benchmark.cpp
… test/unit/alignment/matrix/detail/CMakeLists.txt
… test/unit/alignment/matrix/detail/score_matrix_single_column_test.cpp
… test/unit/alignment/pairwise/edit_distance_unbanded_test_template.hpp
… test/unit/alignment/pairwise/execution/alignment_executor_two_way_test.cpp
… test/unit/alignment/pairwise/pairwise_alignment_single_test_template.hpp