First part of refactoring the alignment to a more optimised version:

Old microbenchmark of global affine alignment without vectorisation:
```
----------------------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------
seqan3_affine_dna4                      912108 ns       901837 ns          724 CUPS=278.322M/s cells=251.001k
seqan2_affine_dna4                      507343 ns       501442 ns         1293 CUPS=500.558M/s cells=251.001k

seqan3_affine_dna4_collection          3780352 ns      3737470 ns          185 CUPS=272.939M/s cells=1020.1k
seqan2_affine_dna4_collection          2544802 ns      2516513 ns          273 CUPS=405.363M/s cells=1020.1k
```

New microbenchmark of global affine alignment without vectorisation:

```
----------------------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------
seqan3_affine_dna4                      508742 ns       501681 ns         1315 CUPS=500.32M/s cells=251.001k
seqan2_affine_dna4                      506287 ns       500708 ns         1280 CUPS=501.292M/s cells=251.001k

seqan3_affine_dna4_collection          2098579 ns      2070642 ns          332 CUPS=492.649M/s cells=1020.1k
seqan2_affine_dna4_collection          2559029 ns      2508209 ns          273 CUPS=406.705M/s cells=1020.1k
```

### TODO: Place here the macobenchmark results.