@rrahn 2020-05-15
Ok, the final vectorised version is now available in #1804.